# src/config/swagger.json
{
  "swagger": "2.0",
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "UnauthorizedError": {
        "description": "User unauthorized"
      },
      "InternalError": {
        "description": "Internal server error"
      }
    }
  },
  "info": {
    "description": "Task Tracker Test App API server.",
    "version": "1.0.0",
    "title": "Task Tracker API"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/auth/signup": {
      "get": {
        "summary": "Get signup route",
        "responses": {
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Create user and save entity to database",
        "operationId": "addUser",
        "responses": {
          "200": {
            "description": "OK, user entity saved"
          },
          "422": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "example": {}
              }
            }
          }
        ]
      }
    },
    "/auth/login": {
      "get": {
        "summary": "Get form for user login",
        "responses": {
          "200": {
            "description": "OK, render login form"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "summary": "User login route",
        "operationId": "login",
        "responses": {
          "200": {
            "description": "OK, returns authorization token"
          },
          "401": {
            "description": "Couldn't authorize user"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "example": {}
              }
            }
          }
        ]
      }
    },
    "/users": {
      "get": {
        "summary": "Get list of all users",
        "responses": {
          "200": {
            "description": "OK, return list of users"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "page",
            "type": "number",
            "description": "Page number"
          }
        ]
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "in": "path",
          "name": "userId",
          "required": true,
          "type": "integer"
        }
      ],
      "get": {
        "summary": "Get user by id",
        "responses": {
          "200": {
            "description": "OK, return user info"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      },
      "put": {
        "summary": "Update user info",
        "responses": {
          "200": {
            "description": "OK, update user and return the entity"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "example": {}
              }
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      },
      "delete": {
        "summary": "Delete user with the given ID.",
        "responses": {
          "200": {
            "description": "OK, user deleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get list of all tasks",
        "responses": {
          "200": {
            "description": "OK, return list of tasks"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          },
          {
            "in": "query",
            "name": "status",
            "type": "string",
            "description": "Show tasks with the given status"
          },
          {
            "in": "query",
            "name": "byUserCreationDate",
            "type": "string",
            "description": "Show tasks by user creation date (newer/older)"
          }
        ]
      }
    },
    "/tasks/add": {
      "post": {
        "summary": "Add task",
        "responses": {
          "200": {
            "description": "OK, task created and saved to database"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "example": {}
              }
            }
          }
        ]
      }
    },
    "/tasks/{taskId}": {
      "parameters": [
        {
          "in": "path",
          "name": "taskId",
          "required": true,
          "type": "integer"
        }
      ],
      "get": {
        "summary": "Get a task by its ID",
        "responses": {
          "200": {
            "description": "OK, return task"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      },
      "put": {
        "summary": "Update task info",
        "responses": {
          "200": {
            "description": "OK, update task and return the entity"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "assignee": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "example": {}
              }
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      },
      "delete": {
        "responses": {
          "200": {
            "description": "OK, task with given ID is deleted"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        },
        "summary": "Delete task with the given ID.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "type": "string",
            "required": true
          }
        ]
      }
    }
  }
}